2013-10-09 21:59 UTC+0200 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * config/globsh.mk     * use $(CP) as workaround for missing soft links in NT shell.       I know about mklink but it needs quite new Windows version       and soft links need special privileges. Anyhow better solution       is highly welcome.
  * config/postinst.hb
    * use hardlinks or copy files if symlinks do not work

  * config/linux/gcc.mk
    * added support for non SH shells - it allows to create Android
      builds in MS-Windows using Android NDK from Google.

+++ config/linux/gcc.mk
@@ -45,13 +45,45 @@ LDLIBS := $(foreach lib,$(HB_USER_LIBS) $(LIBS) $(SYSLIBS),-l$(lib))
 LDFLAGS += $(LIBPATHS)
 
 AR := $(HB_CCPREFIX)ar
-AR_RULE = ( $(AR) $(ARFLAGS) $(HB_AFLAGS) $(HB_USER_AFLAGS) rcs $(LIB_DIR)/$@ $(^F) $(ARSTRIP) ) || ( $(RM) $(LIB_DIR)/$@ && $(FALSE) )
+ifeq ($(HB_SHELL),sh)
+   AR_RULE = ( $(AR) $(ARFLAGS) $(HB_AFLAGS) $(HB_USER_AFLAGS) rcs $(LIB_DIR)/$@ $(^F) $(ARSTRIP) ) || ( $(RM) $(LIB_DIR)/$@ && $(FALSE) )
+else
+
+# NOTE: The empty line directly before 'endef' HAVE TO exist!
+define library_object
+   @$(ECHO) $(ECHOQUOTE)$(subst \,/,$(file))$(ECHOQUOTE) >> __lib__.tmp
+
+endef
+define create_library
+   $(if $(wildcard __lib__.tmp),@$(RM) __lib__.tmp,)
+   $(foreach file,$^,$(library_object))
+   ( $(AR) $(ARFLAGS) $(HB_AFLAGS) $(HB_USER_AFLAGS) rcs $(LIB_DIR)/$@ @__lib__.tmp $(ARSTRIP) ) || ( $(RM) $(subst /,$(DIRSEP),$(LIB_DIR)/$@) && $(FALSE) )
+endef
+AR_RULE = $(create_library)
+
+endif
 
 DY := $(CC)
 DFLAGS += -shared $(LIBPATHS)
 DY_OUT := -o$(subst x,x, )
 DLIBS := $(foreach lib,$(HB_USER_LIBS) $(SYSLIBS),-l$(lib))
 
-DY_RULE = $(DY) $(DFLAGS) -Wl,-soname,$(DYN_NAME_CPT) $(HB_USER_DFLAGS) $(DY_OUT)$(DYN_DIR)/$@ $^ $(DLIBS) $(DYSTRIP) && $(LN) $(@F) $(DYN_FILE_NVR) && $(LN) $(@F) $(DYN_FILE_CPT)
+ifeq ($(HB_SHELL),sh)
+   DY_RULE = $(DY) $(DFLAGS) -Wl,-soname,$(DYN_NAME_CPT) $(HB_USER_DFLAGS) $(DY_OUT)$(DYN_DIR)/$@ $^ $(DLIBS) $(DYSTRIP) && $(LN) $(@F) $(DYN_FILE_NVR) && $(LN) $(@F) $(DYN_FILE_CPT)
+else
+
+# NOTE: The empty line directly before 'endef' HAVE TO exist!
+define dynlib_object
+   @$(ECHO) $(ECHOQUOTE)INPUT($(subst \,/,$(file)))$(ECHOQUOTE) >> __dyn__.tmp
+
+endef
+define create_dynlib
+   $(if $(wildcard __dyn__.tmp),@$(RM) __dyn__.tmp,)
+   $(foreach file,$^,$(dynlib_object))
+   $(DY) $(DFLAGS) -Wl,-soname,$(DYN_NAME_CPT) $(HB_USER_DFLAGS) $(DY_OUT)$(DYN_DIR)/$@ __dyn__.tmp $(DLIBS) $(DYSTRIP) && $(LN) $(subst /,$(DIRSEP),$(DYN_DIR)/$@ $(DYN_FILE_NVR)) && $(LN) $(subst /,$(DIRSEP),$(DYN_DIR)/$@ $(DYN_FILE_CPT))
+endef
+DY_RULE = $(create_dynlib)
+
+endif
 
 include $(TOP)$(ROOT)config/rules.mk
